services/users/requirements.txt
flake8==3.7.9
Flask-cors
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
marshmallow==3.5.1
psycopg2-binary==2.8.4
pyjwt==1.7.1
pytest==5.3.5
pytest-cov==2.8.1
python-dotenv==0.11.0
